--- rust-client/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod common;
